fixed some minor issue

+++ AIMS_BD_IATI.website/Modules/IATIImport/Dashboard/DashboardController.js
@@ -112,7 +112,7 @@ angular.module('iatiDataImporter').filter('sumByKey', function() {
 
             var sum = 0;
             for (var i = data.length - 1; i >= 0; i--) {
-                sum += parseInt(data[i][key]);
+                sum += parseFloat(data[i][key]);
             }
 
             return sum;
